add loginform and style it, got to go with the backend
