+++ frontend/src/Components/Buttons.css
@@ -5,7 +5,7 @@
   border: none;
   border-radius: 6px;
   cursor: pointer;
-  margin: 3px 0;
+  margin: 5px 0;
   width: 100%;
 }
