+++ frontend/src/Components/Footer.css
@@ -7,8 +7,8 @@
   justify-content: space-between;
   align-items: center;
   padding: 5px;
-  background-color: #cbcbcb;
-  color: rgb(71, 70, 70);
+  background-color: #303030;
+  color: rgb(255, 255, 255);
   width: 100%;
 }
