+++ frontend/src/Components/Header.css
@@ -7,10 +7,6 @@
   font-size: 1.5rem;
   width: 100%;
   margin-bottom: 20px;
-  background-image: url("/keys-wall.jpg");
-  background-attachment: fixed;
-  background-position: center;
-  background-size: 100%;
 }
 
 .header-container h1 {
@@ -24,15 +20,13 @@
 
 /* TABLET SCREENS min 745px ------------------------------------------------------------ */
 
-@media (min-width:745px){ 
+@media (min-width: 745px) {
   .header-container {
     min-height: 300px;
   }
 
   .header-container h1 {
-
     font-size: 4rem;
     backdrop-filter: blur(5px);
-
   }
 }
