+++ frontend/src/Components/LoginForm.css
@@ -1,4 +1,4 @@
-.sign-in-container {
+.login-container {
   z-index: 10;
   position: absolute;
   background-color: rgba(255, 255, 255, 0.837);
@@ -9,7 +9,6 @@
   align-items: center;
   padding: 10px;
   width: 100%;
-  height: 170px;
   backdrop-filter: blur(2px);
   border-radius: 6px;
   top: 0;
@@ -17,7 +16,7 @@
   border: 2px solid white;
 }
 
-.sign-in-container > div {
+.login-container > div {
   position: relative;
   width: 100%;
   display: flex;
@@ -44,7 +43,7 @@
   color: grey;
 }
 
-.sign-in-container input {
+.input-field > input {
   padding: 8px 16px;
   padding-left: 1px;
   border: none;
@@ -55,7 +54,13 @@
   margin: 3px 0;
 }
 
-/* TABLET SCREENS min 745px ------------------------------------------------------------ */
+small {
+  font-size: 0.7rem;
+  color: black;
+  align-self: flex-start;
+}
+
+ TABLET SCREENS min 745px ------------------------------------------------------------ */
 
 @media (min-width: 745px) {
 }
